requirements.txt
absl-py==2.1.0
aenum==3.1.15
affine==2.4.0
albumentations==1.3.0
astrapy==0.3.3
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backoff==2.2.1
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
cassandra-driver==3.28.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
executing==2.1.0
filelock==3.16.0
filetype==1.2.0
firebase==4.0.1
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
gast==0.6.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gql==3.4.1
graphql-core==3.2.3
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==20.1.0
h5py==3.11.0
httplib2==0.22.0
idna==3.7
imageio==2.35.1
importlib-metadata==4.11.4
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.32
namex==0.0.8
neo4j==5.9.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.3.3
pyparsing==3.1.4
PyPDF2==2.11.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
qudida==0.0.4
rasterio==1.4.1
redis==4.5.5
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
roboflow==1.1.48
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
supervision==0.24.0
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.0.196
ultralytics-thop==2.0.6
uritemplate==4.1.1
urllib3==2.2.2
vercel==0.2.1
watchdog==5.0.2
wcwidth==0.2.13
Werkzeug==3.0.4
wget==3.2
wrapt==1.16.0
yarl==1.9.2
zipp==3.8.0
